Allow overriding platform specifica via platform def file
The code in package.def will include the following files if they
exist, in this order:
  $(PKG_ROOT)/$(BUILD_DISTRIBUTION).def
  $(PKG_ROOT)/$(BUILD_PLATFORM).def

Eg. zimbra/os-requirements/Makefile will look into
  zimbra/os-requirements/UBUNTU.def
  zimbra/os-requirements/UBUNTU14_64.def

The main purpose is to pull out platform specific variables to
these files and allow a less intrusive porting to a new
distribution/platform (eg. DEBIAN).

This is not supposed a solution for all platform specific parts
in the Makefiles, just the most common ones.