ZCS-5790:login gets 502 when one of MBS servers is down
Problem
In general, Nginx chooses the initial upstream MBS (where the login page is fetched) based on the client's source IP address.  When the chosen upstream MBS is down, Nginx is supposed to switch to other MBS; it works when the zimbraMailUrl is set to the default value '/'.  But, if the zimbraMailUrl is set to other than '/', such as '/zimbra' the former default value of the 'zimbraMailUrl', the Nginx does not failover the running MBS but returns 502 error.

This bug was introduced by the Bugzilla 79621.  The bug fix description said:
```
Added a check for HTTP POST requests to terminate the connection if the upstream server is down and if the request is not a auth request.
Web client auth request are HTTP POST request with URI '/' and can also include URL prefix from 'zimbraMailURL'.
```
But in fact, if the 'zimbraMailURL' was other than '/', the code compared not the *prefix* of the HTTP Post but the *entire* part of the request with the 'zimbraMailURL'.   As a result, the nginx coudn't detect the down of the upstream server.

Who needs this fix:
The system operating with multiple MBS servers and haveing a configuration of "zimbraMailURL" other than "/".